+++ services/sso/src/controllers/auth/registerController.ts
@@ -1,8 +1,10 @@
 import { Context, Middleware } from 'koa'
 import { UserRegister } from '../../schemas/auth/registerSchema'
-import { client } from '../../db/client'
+import db from '../../db/knexClient'
 import { hashPassword } from '../../utils/passwordHash'
 import { generateId } from '../../utils/cuidGenerator'
+import { userManager } from '../../db/managers/userManager'
+import { DuplicateError } from '../../utils/errors'
 
 export const registerController: Middleware = async (ctx: Context) => {
   const { dni, email, name, password, itineraryId }: UserRegister =
@@ -11,20 +13,26 @@ export const registerController: Middleware = async (ctx: Context) => {
   const hashedPassword = await hashPassword(password)
   const id = generateId()
 
-  const query = {
-    text: 'INSERT INTO "user"(id, dni, email, name, password, user_meta, itinerary_id) VALUES($1, $2, $3, $4, $5, $6, $7)',
-    values: [
-      id,
-      dniToUpperCase,
-      email,
-      name,
-      hashedPassword,
-      '{}',
-      itineraryId,
-    ],
+  const dniNotValid = await userManager.findByDni(dniToUpperCase, {
+    fields: ['dni'],
+  })
+  const emailNotValid = await userManager.getUser(email, {
+    fields: ['email'],
+  })
+
+  if (dniNotValid || emailNotValid) {
+    throw new DuplicateError('email or dni already exists')
   }
 
-  await client.query(query)
+  await db('user').insert({
+    id,
+    dni: dniToUpperCase,
+    email,
+    name,
+    password: hashedPassword,
+    user_meta: '{}',
+    itinerary_id: itineraryId,
+  })
 
   ctx.status = 200
   ctx.body = { id }